backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sqlite3-binary
